requirement.txt
python-dotenv
google-generativeai
langchain
faiss-cpu
pypdf
langchain_community
langchain_google_genai